+++ src/commands/unknown_command.rs
@@ -1,8 +1,8 @@
 use crate::{
-    commands::{Command, HelpMessage, COMMANDS},
+    HandlerResult,
+    commands::{COMMANDS, Command, HelpMessage},
     consts::{BOT_NAME, MIN_SIMILARITY_SCORE},
     utils::trace_skip_all,
-    HandlerResult,
 };
 use macro_rules_attribute::apply;
 use rust_fuzzy_search::fuzzy_search_best_n;
@@ -16,59 +16,52 @@ pub async fn unknown_command(bot: Bot, msg: Message) -> HandlerResult {
         return Ok(());
     };
 
-    let Some(mut command) = text.strip_prefix('/') else {
+    let command = text.split_whitespace().next().unwrap_or("");
+
+    let Some(mut cmd_name) = text.strip_prefix('/') else {
         return Ok(());
     };
 
-    command = command.split_whitespace().next().unwrap_or("");
+    cmd_name = cmd_name
+        .strip_suffix(&format!("@{BOT_NAME}"))
+        .unwrap_or(cmd_name);
 
-    let Some(command) = command.strip_suffix(&format!("@{BOT_NAME}")) else {
-        return Ok(());
-    };
+    let available_cmd_names: Vec<&str> = COMMANDS.iter().map(String::as_ref).collect();
 
-    let available_commands: Vec<&str> = COMMANDS.iter().map(String::as_ref).collect();
-
-    if available_commands.contains(&command) {
-        let help_message = Command::from_str(command)
-            .unwrap_or_else(|_| panic!("Command /{} should exist.", command))
+    if available_cmd_names.contains(&cmd_name) {
+        let help_message = Command::from_str(cmd_name)
+            .unwrap_or_else(|_| panic!("Command {command} should exist."))
             .help_message();
 
         bot.send_message(
             msg.chat.id,
-            format!(
-                "Invalid usage of command: /{}.\n\n{}",
-                command, help_message
-            ),
+            format!("Invalid usage of command: {command}.\n\n{help_message}",),
         )
         .await?;
-    } else if available_commands.contains(&command.to_lowercase().as_str()) {
+    } else if available_cmd_names.contains(&cmd_name.to_lowercase().as_str()) {
         bot.send_message(
             msg.chat.id,
             format!(
-                "Unknown command: {}.\nDid you mean: /{}?",
-                text,
-                command.to_lowercase()
+                "Unknown command: {text}.\nDid you mean: /{}?",
+                cmd_name.to_lowercase()
             ),
         )
         .await?;
     } else {
-        let (best_match, best_score) = fuzzy_search_best_n(command, &available_commands, 1)[0];
+        let (best_match, best_score) = fuzzy_search_best_n(cmd_name, &available_cmd_names, 1)[0];
 
         tracing::debug!(
-            "Input command: {}, best match: {}, score: {}.",
-            command,
-            best_match,
-            best_score
+            "Input command: {cmd_name}, best match: {best_match}, score: {best_score}."
         );
 
         if best_score >= MIN_SIMILARITY_SCORE {
             bot.send_message(
                 msg.chat.id,
-                format!("Unknown command: {}.\nDid you mean: /{}?", text, best_match),
+                format!("Unknown command: {text}.\nDid you mean: /{best_match}?"),
             )
             .await?;
         } else {
-            bot.send_message(msg.chat.id, format!("Unknown command: {}.", text))
+            bot.send_message(msg.chat.id, format!("Unknown command: {text}."))
                 .await?;
         }
     }